App/python-server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = "Eprice App Python Server"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "asyncpg==0.30.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose==3.4.0",
    "uvicorn==0.34.2",
    "pydantic[email]==2.11.4",
    "dotenv==0.9.9",
    "httpx==0.28.1",
    "asyncio==3.4.3",
    "apscheduler==3.11.0",
    "requests==2.32.3",
    "fastapi-mail==1.4.2",
]
